refactor(web-viewer): move status-filter.ts to views/ to align with zone pin
The file was pinned to the web-viewer zone but lived in
components/prd-tree/, creating a location/zone mismatch.
Moves it to views/ where other web-viewer zone files reside.

Updates all consumers (prd-tree.ts, index.ts barrel,
use-persistent-filter.ts hook, three test files) to import
from the new path. Zone pin in .n-dx.json and ZONES.md manifest
updated to match the new location.
